calori, double yapıldı.

+++ VeganFit.Entities/OptionalProduct.cs
@@ -13,7 +13,7 @@ namespace VeganFit.Entities
     public class OptionalProduct : BaseEntity
     {
         public string ProductName { get; set; }
-        public int Calori { get; set; }
+        public double Calori { get; set; }
         public string? Serving { get; set; }
         public byte[] Picture { get; set; }
         public virtual ICollection<Data> Datas { get; set; }
